preliminary version of Plink pipeline, can only run basic hap assoc

+++ pycmm/template.py
@@ -6,7 +6,6 @@ import subprocess
 import inspect
 import fileinput
 import gc
-import tempfile
 import string
 from random import choice
 from os.path import join as join_path
@@ -33,15 +32,9 @@ class pyCMMBase(object):
     def get_raw_repr(self):
         return "Not yet implemented"
 
-    @property
-    def local_scratch_dir(self):
-        return tempfile.mkdtemp()
-
-    @property
-    def local_tmp_file(self):
+    def get_tmp_file_name(self):
         chars = string.ascii_letters
-        file_name = "tmp" + ''.join([choice(chars) for i in range(6)])
-        return join_path(self.local_scratch_dir, file_name)
+        return "tmp_" + ''.join([choice(chars) for i in range(6)])
 
     def remove_dir(self, dir_name):
         if os.path.exists(dir_name):
@@ -64,7 +57,8 @@ class pyCMMBase(object):
                 fout.write(line)
 
     def copy_file(self, src, dst):
-        self.delete_file(dst)
+        if not os.path.isdir(dst):
+            self.delete_file(dst)
         if os.path.islink(src):
             linkto = os.readlink(src)
             os.symlink(linkto, dst)
